requirements.txt
Flask
authlib
requests
python-dotenv